Follow-Up (módulo completo): fila do dia, esgotados, KPIs, cobertura e config
O 8º sistema ganha corpo: a rota /follow-up entrega a fila um-por-vez no
modelo do Modo Volume (ação primária no canal que a régua manda, desfecho
de 1 clique que agenda o próximo toque via tarefa comum — nunca escrevendo
proximo_followup direto), a visão de Esgotados para a decisão humana
(reativar/descartar, nunca auto-perdido), os KPIs da curva de resposta por
tentativa (recharts só no bundle lazy) e a Cobertura do time para a gestão.

- Mensagem do toque: convenção "Régua N" na biblioteca de templates, com
  fallback G.P.V.A. embutido por fase da régua (≤4 linhas, personalizado).
- Configurações ganham a aba Follow-Up: editor draft+Salvar da régua em
  gestao_config (teto, matriz de gaps 3×N, toques por ligação,
  multiplicadores de etapa, SLA + opt-in da devolução automática).
- Migration: LEAST(...)::int na MV followup_tentativa_mensal — row_number é
  bigint e as RPCs de curva declaram RETURNS TABLE (tentativa int, ...);
  sem o cast, RETURN QUERY falhava com 42804 para qualquer caller.
- Testes: 20 de banco (contador com colapso de sessão, fila, esgotar/
  reativar, nav v3, devolução por SLA, MV + RPCs), 24 de unidade novos;
  metrics-camada atualizado para as 5 MVs do refresh_all.

+++ supabase/migrations/20260827130000_followup_regua.sql
@@ -504,7 +504,9 @@ toques AS (
 ),
 avaliados AS (
   SELECT t.lead_id, t.quando, t.corretor_id,
-         LEAST(t.tentativa, 20) AS tentativa,
+         -- ::int porque row_number() é bigint e as RPCs de curva declaram
+         -- RETURNS TABLE (tentativa int, ...) — RETURN QUERY exige tipo exato.
+         LEAST(t.tentativa, 20)::int AS tentativa,
          EXISTS (
            SELECT 1 FROM public.interacoes r
            WHERE r.lead_id = t.lead_id
